Rename inference_pt to inference
