+++ sockeye/checkpoint_decoder.py
@@ -27,7 +27,7 @@ import torch
 from . import constants as C
 from . import data_io
 from . import evaluate
-from . import inference_pt
+from . import inference
 from . import model_pt
 from . import utils
 from . import vocab
