+++ test/unit/test_output_handler.py
@@ -16,7 +16,7 @@ import io
 import pytest
 
 import sockeye.output_handler
-from sockeye.inference_pt import TranslatorInput, TranslatorOutput
+from sockeye.inference import TranslatorInput, TranslatorOutput
 
 stream_handler_tests = [(sockeye.output_handler.StringOutputHandler(io.StringIO()),
                          TranslatorInput(sentence_id=0, tokens=[], factors=[], constraints=[]),
